+++ forge/src/main/java/net/blay09/mods/balm/forge/stats/ForgeBalmStats.java
@@ -1,8 +1,8 @@
 package net.blay09.mods.balm.forge.stats;
 
 import net.blay09.mods.balm.api.stats.BalmStats;
-import net.minecraft.core.Registry;
-import net.minecraft.core.registries.BuiltInRegistries;
+import net.blay09.mods.balm.forge.DeferredRegisters;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.stats.StatFormatter;
 import net.minecraft.stats.Stats;
@@ -22,6 +22,9 @@ public class ForgeBalmStats implements BalmStats {
     @Override
     public void registerCustomStat(ResourceLocation identifier) {
         getRegistrations(identifier.getNamespace()).customStats.add(identifier);
+
+        final var registry = DeferredRegisters.get(Registries.CUSTOM_STAT, identifier.getNamespace());
+        registry.register(identifier.getPath(), () -> identifier);
     }
 
     public void register(String modId, IEventBus eventBus) {
@@ -37,10 +40,7 @@ public class ForgeBalmStats implements BalmStats {
 
         @SubscribeEvent
         public void commonSetup(FMLCommonSetupEvent event) {
-            event.enqueueWork(() -> customStats.forEach(it -> {
-                Registry.register(BuiltInRegistries.CUSTOM_STAT, it.getPath(), it);
-                Stats.CUSTOM.get(it, StatFormatter.DEFAULT);
-            }));
+            event.enqueueWork(() -> customStats.forEach(it -> Stats.CUSTOM.get(it, StatFormatter.DEFAULT)));
         }
     }
 }